requirements.txt
astroid==2.4.2
bcrypt==3.2.0
click==7.1.2
crypto==1.4.1
cycler==0.10.0
Flask==1.1.2
Flask-Session==0.3.2
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
Naked==0.1.31
numpy==1.19.2
Pillow==8.0.0
pylint==2.6.0
pyopenssl==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.3.1
redis~=3.5.3
requests==2.24.0
shellescape==3.8.1
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
flask-cors
flask-jwt-extended==3.25.0
Flask-HAL==1.0.4
flask
python-dotenv
requests
authlib
six